pyproject.toml
[project]
name = "reduce-image-size"
version = "0.1.0"
description = "Resize images"
readme = "README.md"
requires-python = ">=3.12, <=3.13"
authors = [
    {name = "sergejyit@gmail.com"},
]
dependencies = [
    "Pillow",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.ruff]
target-version = "py312"
lint.select = [
    "ALL",
]
lint.ignore = ["D103", "D100", "D203", "D213"]
